Add room action tests
