+++ src/reducers/root.js
@@ -11,7 +11,7 @@ import {
   past
 } from './prepareData.js';
 
-import { depth, width, height, numberOfPeople } from './size.js';
+import { depth, width, height, numberOfPeople } from './room.reducer.js';
 import { appConfig } from './config.js';
 
 export default (state = {}, action) => ({
